Still moving endpoints under their own routers
